chore(integration): update integration tests
